fix: consolidate JWT authentication to single source of truth
- Centralized JWT verification and signing utilities in middleware/auth.ts
- Removed duplicate JWT functions from auth.ts routes
- Fixed JWT_SECRET inconsistency between auth route and middleware
- Added proper JWT_SECRET fallback for development environment
- Created reusable optionalAuth middleware to replace duplicated code
- All JWT operations now use consistent secret and verification logic
